split GPSListener class
