+++ app/src/main/java/com/guardiansofgalakddy/lvlmonitor/MonitorActivity.java
@@ -30,6 +30,7 @@ import com.google.android.gms.maps.model.BitmapDescriptorFactory;
 import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.maps.model.MarkerOptions;
 
+import com.guardiansofgalakddy.lvlmonitor.Onegold.Map.GPSListener;
 import com.guardiansofgalakddy.lvlmonitor.junhwa.BLEScanner;
 
 /* Google Map 관련 코드 - 천우진
@@ -39,9 +40,8 @@ import com.guardiansofgalakddy.lvlmonitor.junhwa.BLEScanner;
 
 public class MonitorActivity extends AppCompatActivity {
     /* Map object */
-    SupportMapFragment mapFragment;
-    GoogleMap map;
-    MarkerOptions myLocationMarker;
+    private SupportMapFragment mapFragment;
+    private GPSListener gpsListener;
 
 
     private RecyclerAdapter adapter;
@@ -130,13 +130,16 @@ public class MonitorActivity extends AppCompatActivity {
 
     /* Google Map 처음 설정 */
     private void initGoogleMap() {
+        /* get GPSListener object */
+        gpsListener = GPSListener.getInstance();
         /* Google Map Fragment 등록 */
         mapFragment = (SupportMapFragment) getSupportFragmentManager().findFragmentById(R.id.map_monitor);
         mapFragment.getMapAsync(new OnMapReadyCallback() {
             @Override
             public void onMapReady(GoogleMap googleMap) {
-                map = googleMap;
+                GoogleMap map = googleMap;
                 map.setMyLocationEnabled(true);
+                gpsListener.setMap(map);
                 /* Google Map location setting */
                 startLocationService();
             }
@@ -159,15 +162,8 @@ public class MonitorActivity extends AppCompatActivity {
                 Toast.makeText(getApplicationContext(), "GPS 권한이 없거나 위치 기능이 꺼져있습니다.", Toast.LENGTH_LONG).show();
                 finish();
             }
-            // 위치 정보 획득
-            Location location = manager.getLastKnownLocation(LocationManager.NETWORK_PROVIDER);
-            if (location != null) {
-                double latitude = location.getLatitude();
-                double longitude = location.getLongitude();
-                Log.d("startLocationService: ", latitude + ", " + longitude);
-            }
+
             // GPSListener register
-            GPSListener gpsListener = new GPSListener();
             long minTime = 10000;
             float minDistance = 0;
 
@@ -178,117 +174,12 @@ public class MonitorActivity extends AppCompatActivity {
         }
     }
 
-    /* GPS Listener */
-    public class GPSListener implements LocationListener {
-        private static final int TWO_MINUTES = 1000 * 60 * 2;
-        Location currentLocation;
-
-        @Override
-        public void onLocationChanged(Location location) {
-            Double latitude;
-            Double longitude;
-
-            // check best location provider
-            if (isBetterLocation(location, currentLocation)) {
-                currentLocation = location;
-            }
-            latitude = currentLocation.getLatitude();
-            longitude = currentLocation.getLongitude();
-            Log.d("onLocationChanged: ", latitude + ", " + longitude + ", " + currentLocation.getProvider());
-            // move camera to current location
-            showCurrentLocation(latitude, longitude);
-        }
-
-        // check newLocation better than currentLocation
-        public boolean isBetterLocation(Location location, Location currentBestLocation) {
-            if (currentBestLocation == null) {
-                return true;
-            }
-
-            long timeDelta = location.getTime() - currentBestLocation.getTime();
-            boolean isSignificatlyNewer = timeDelta > TWO_MINUTES;
-            boolean isSignificantlyOlder = timeDelta < -TWO_MINUTES;
-            boolean isNewer = timeDelta > 0;
-
-            if (isSignificatlyNewer) {
-                return true;
-            } else if (isSignificantlyOlder) {
-                return false;
-            }
-
-            int accuracyDelta = (int) (location.getAccuracy() - currentBestLocation.getAccuracy());
-            boolean isLessAccurate = accuracyDelta > 0;
-            boolean isMoreAccurate = accuracyDelta < 0;
-            boolean isSignificantlyLessAccurate = accuracyDelta > 200;
-
-            boolean isFromSameProvider = isSameProvider(location.getProvider(), currentBestLocation.getProvider());
-
-            if (isMoreAccurate) {
-                return true;
-            } else if (isNewer && !isLessAccurate) {
-                return true;
-            } else if (isNewer && !isSignificantlyLessAccurate && isFromSameProvider) {
-                return true;
-            }
-            return false;
-        }
-
-        // check same location Provider
-        private boolean isSameProvider(String provider1, String provider2) {
-            if (provider1 == null) {
-                return provider2 == null;
-            }
-            return provider1.equals(provider2);
-        }
-
-        // 현재 위치로 Google Map 이동
-        private void showCurrentLocation(Double latitude, Double longitude) {
-            LatLng curPoint = new LatLng(latitude, longitude);
-            map.animateCamera(CameraUpdateFactory.newLatLngZoom(curPoint, 15));
-            Log.d("showCurrentLocation : ", latitude + ", " + longitude);
-        }
-
-        // Show my location Marker
-        private void showMyLocationMarker(LatLng curPoint) {
-            if (myLocationMarker == null) {
-                myLocationMarker = new MarkerOptions()
-                        .position(curPoint)
-                        .icon(BitmapDescriptorFactory.defaultMarker(BitmapDescriptorFactory.HUE_ORANGE));
-                map.addMarker(myLocationMarker);
-            } else {
-                myLocationMarker.position(curPoint);
-            }
-        }
-
-        @Override
-        public void onStatusChanged(String s, int i, Bundle bundle) {
-
-        }
-
-        @Override
-        public void onProviderEnabled(String s) {
-
-        }
-
-        @Override
-        public void onProviderDisabled(String s) {
-
-        }
-    }
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-        if (map != null) {
-            map.setMyLocationEnabled(true);
-        }
-    }
 
     @Override
     protected void onDestroy() {
         super.onDestroy();
-        if (map != null) {
-            map.setMyLocationEnabled(false);
+        if (gpsListener != null) {
+            gpsListener.getMap().setMyLocationEnabled(false);
         }
     }
 }
