+++ app/src/main/res/raw/third_party_license_metadata
@@ -1,43 +1,35 @@
-0:46 material
-48:46 mediarouter
-96:46 legacy-support-v4
-144:46 recyclerview
-192:46 palette
-240:46 media2
-288:46 media
-336:46 transition
-384:46 cardview
-432:42 AutoPermissions
-1964:46 kotlin-stdlib-jdk7
-2012:46 kotlin-stdlib
-572:46 lint-gradle
-620:46 builder
-668:46 lint
-716:46 lint-checks
-764:46 lint-api
-812:46 manifest-merger
-860:46 sdk-common
-908:46 lint-gradle-api
-956:46 gradle-api
-1004:46 builder-model
-1052:46 intellij-core
-1100:46 kotlin-compiler
-1148:46 uast
-1196:46 groovy-all
-1244:46 kotlin-reflect
-1292:46 tracker
-1340:46 shared
-1388:46 builder-test-api
-1436:46 ddmlib
-1484:46 sdklib
-1532:46 layoutlib-api
-1580:46 dvlib
-1628:46 repository
-1676:46 common
-1724:46 kotlin-stdlib-jdk8
-1772:46 apkzlib
-1820:46 protos
-1868:46 apksig
-1916:46 annotations
-2060:46 error_prone_annotations
-2108:46 kotlin-stdlib-common
+0:46 mediarouter
+48:46 palette
+96:46 media2
+144:42 AutoPermissions
+188:46 lint-gradle
+236:46 builder
+284:46 lint
+332:46 lint-checks
+380:46 lint-api
+428:46 manifest-merger
+476:46 sdk-common
+524:46 lint-gradle-api
+572:46 gradle-api
+620:46 builder-model
+668:46 intellij-core
+716:46 kotlin-compiler
+764:46 uast
+812:46 kotlin-reflect
+860:46 tracker
+908:46 shared
+956:46 builder-test-api
+1004:46 ddmlib
+1052:46 sdklib
+1100:46 layoutlib-api
+1148:46 dvlib
+1196:46 repository
+1244:46 common
+1292:46 kotlin-stdlib-jdk8
+1340:46 apkzlib
+1388:46 protos
+1436:46 apksig
+1484:46 annotations
+1532:46 kotlin-stdlib-jdk7
+1580:46 kotlin-stdlib
+1628:46 kotlin-stdlib-common
