+++ app/src/main/res/raw/third_party_licenses
@@ -1,17 +1,7 @@
 http://www.apache.org/licenses/LICENSE-2.0.txt
 http://www.apache.org/licenses/LICENSE-2.0.txt
 http://www.apache.org/licenses/LICENSE-2.0.txt
-http://www.apache.org/licenses/LICENSE-2.0.txt
-http://www.apache.org/licenses/LICENSE-2.0.txt
-http://www.apache.org/licenses/LICENSE-2.0.txt
-http://www.apache.org/licenses/LICENSE-2.0.txt
-http://www.apache.org/licenses/LICENSE-2.0.txt
-http://www.apache.org/licenses/LICENSE-2.0.txt
 https://api.github.com/licenses/apache-2.0
-http://www.apache.org/licenses/LICENSE-2.0.txt
-http://www.apache.org/licenses/LICENSE-2.0.txt
-http://www.apache.org/licenses/LICENSE-2.0.txt
-http://www.apache.org/licenses/LICENSE-2.0.txt
 http://www.apache.org/licenses/LICENSE-2.0.txt
 http://www.apache.org/licenses/LICENSE-2.0.txt
 http://www.apache.org/licenses/LICENSE-2.0.txt
